Remove maven its dependencies and update readme file
