+++ src/Main.java
@@ -1,0 +1,8 @@
+import controller.UserController;
+
+public class Main {
+    public static void main(String[] args) {
+        UserController userController = new UserController();
+        userController.initialize();
+    }
+}
