+++ src/scripts/check_if_country_exists.sh
