+++ src/scripts/fetch-patient-profile.sh
