+++ src/scripts/get-country-life-expectancy.sh
