+++ src/scripts/get_profile.sh
