+++ src/scripts/init_store.sh
