+++ src/scripts/register_patient.sh
