+++ src/scripts/verify_user.sh
