+++ src/utils/BashRunner.java
@@ -1,4 +1,4 @@
-package com.prognosis.cli.utils;
+package utils;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
