+++ src/view/PatientView.java
@@ -1,8 +1,8 @@
-package com.prognosis.cli.view;
+package view;
 
 import java.util.Scanner;
 
-import com.prognosis.cli.controller.UserController;
+import controller.UserController;
 
 public class PatientView implements UserView {
     UserController userController = new UserController();
@@ -29,5 +29,23 @@ public class PatientView implements UserView {
     @Override
     public void displayErrorMessage(String message) {
         System.out.println(message);
+    }
+
+    @Override
+    public void welcomeMenu() {
+        System.out.println("Welcome Admin!");
+        System.out.println("1. Login");
+        System.out.println("2. Register");
+        System.out.println("3. Exit");
+        do {
+            UserController userController = new UserController();
+            int choice = Integer.parseInt(System.console().readLine());
+            switch (choice) {
+                case 1 -> userController.handleLogin();
+                case 2 -> userController.registerPatientProfile();
+                case 3 -> userController.logout();
+                default -> System.out.println("Invalid choice");
+            }
+        } while (true);
     }
 }
