+++ src/view/UserView.java
@@ -1,7 +1,8 @@
-package com.prognosis.cli.view;
+package view;
 
 public interface UserView {
 
     public void displayMenu();
     public void displayErrorMessage(String message);
+    public void welcomeMenu();
 }
